Fixed bug ' Cannot parse the constant number' in activity file.  Activity Speed Restriction X, and Z, position were ints, s/b floats.
git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@887 a9f3cfff-6c41-7446-b0ca-052bd761e6eb

+++ Source/MSTS/ACTFile.cs
@@ -913,8 +913,8 @@ namespace MSTS {
     public class Position {
         int TileX;
         int TileZ;
-        int X;
-        int Z;
+        float X;
+        float Z;
 
         public Position(int tileX, int tileZ, int x, int z) {
             TileX = tileX;
@@ -927,8 +927,8 @@ namespace MSTS {
             stf.MustMatch("(");
             TileX = stf.ReadInt(STFReader.UNITS.None, null);
             TileZ = stf.ReadInt(STFReader.UNITS.None, null);
-            X = stf.ReadInt(STFReader.UNITS.None, null);
-            Z = stf.ReadInt(STFReader.UNITS.None, null);
+            X = stf.ReadFloat(STFReader.UNITS.None, null);
+            Z = stf.ReadFloat(STFReader.UNITS.None, null);
             stf.MustMatch(")");
         }
     }
